refactor: wire callers to lint module, remove evaluator bootstrap
- file-evaluation.ts: replace createEvaluator/evaluate with runLint();
  drop accuracy-evaluator-specific criteria validation block; remove
  ConfigError, Type, and searchProvider from imports/destructuring
- tool-handlers.ts: replace inline composeSystemPrompt + runPromptStructured
  call with runLintMerged(); import LintRuleCall from src/lint instead of
  lint-prompt; remove buildMergedCheckLLMSchema, MergedCheckLLMResult,
  composeSystemPrompt, and buildLintSystemPrompt imports
- prompt-builder.ts: remove buildLintSystemPrompt (moved to src/lint/),
  remove mergeRulesForLint and LintRuleCall imports from lint-prompt
- src/index.ts: remove evaluator bootstrap side-effect import

Bundle size: 215 KB → 201 KB (evaluator registry no longer loaded at startup)
