+++ src/agent/prompt-builder.ts
@@ -1,5 +1,4 @@
 import type { MatchedRuleUnit } from './rule-units';
-import { mergeRulesForLint, type LintRuleCall } from './lint-prompt';
 
 export interface BuildAgentSystemPromptParams {
   workspaceRoot: string;
@@ -37,15 +36,6 @@ function formatMatchedRuleUnits(
       return `- ${file}\n${renderedUnits}`;
     })
     .join('\n');
-}
-
-export function buildLintSystemPrompt(ruleCalls: LintRuleCall[]): string {
-  return [
-    'Review the file against all of the following source-backed rules.',
-    'Keep findings attributed to the exact ruleSource that each issue belongs to.',
-    '',
-    mergeRulesForLint(ruleCalls),
-  ].join('\n').trim();
 }
 
 export function buildSubAgentSystemPrompt(params: {
